+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.AppTheme">
 
         <activity
-            android:name="com.adamgeraghty.app.MainActivity"
+            android:name="com.adamgeraghty.scorecard.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.AppTheme.NoActionBar"
             android:windowSoftInputMode="adjustResize">
