+++ app/src/main/java/com/adamgeraghty/scorecard/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.adamgeraghty.app
+package com.adamgeraghty.scorecard
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -10,9 +10,9 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.DisposableEffect
 import androidx.compose.ui.Modifier
 import androidx.navigation.compose.rememberNavController
-import com.adamgeraghty.app.theme.ScorecardMaterialTheme
-import com.adamgeraghty.app.ui.navigation.ScorecardNavGraph
-import com.adamgeraghty.app.ui.navigation.Screens
+import com.adamgeraghty.scorecard.theme.ScorecardMaterialTheme
+import com.adamgeraghty.scorecard.ui.navigation.ScorecardNavGraph
+import com.adamgeraghty.scorecard.ui.navigation.Screens
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
