+++ app/src/main/java/com/adamgeraghty/scorecard/di/AppModule.kt
@@ -1,4 +1,4 @@
-package com.adamgeraghty.app.di
+package com.adamgeraghty.scorecard.di
 
 import dagger.Module
 import dagger.hilt.InstallIn
