+++ app/src/main/java/com/adamgeraghty/scorecard/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.adamgeraghty.app.theme
+package com.adamgeraghty.scorecard.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
