+++ app/src/main/java/com/adamgeraghty/scorecard/ui/navigation/NavGraph.kt
@@ -1,4 +1,4 @@
-package com.adamgeraghty.app.ui.navigation
+package com.adamgeraghty.scorecard.ui.navigation
 
 import androidx.compose.animation.core.FastOutSlowInEasing
 import androidx.compose.animation.core.tween
@@ -9,7 +9,7 @@ import androidx.compose.runtime.Composable
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
-import com.adamgeraghty.app.ui.screens.HomeScreen
+import com.adamgeraghty.scorecard.ui.screens.HomeScreen
 
 @Composable
 fun ScorecardNavGraph(
